# knowledge-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/knowledge/topics": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "listTopics",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TopicsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/topics/{slug}/stats": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "topicStats",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "top_tag_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TopicStatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/tags": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "listTags",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/suggest_topic": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "suggestTopic",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/sources": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "listSources",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/scopes": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "listScopes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ScopesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/inbox": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "listInbox",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InboxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/find_duplicates": {
      "get": {
        "tags": ["knowledge-discovery-controller"],
        "operationId": "findDuplicates",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "default": 0.85
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicatesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/audit": {
      "get": {
        "tags": ["knowledge-audit-controller"],
        "operationId": "listAudit",
        "parameters": [
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TopicResponse": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "noteCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastCapturedAt": {
            "type": ["string", "null"]
          },
          "description": {
            "type": ["string", "null"]
          }
        },
        "required": ["noteCount", "slug"]
      },
      "TopicsResponse": {
        "type": "object",
        "properties": {
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicResponse"
            }
          }
        },
        "required": ["topics"]
      },
      "TagResponse": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "lastUsedAt": {
            "type": ["string", "null"]
          }
        },
        "required": ["count", "tag"]
      },
      "TopicStatsResponse": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "noteCount": {
            "type": "integer",
            "format": "int32"
          },
          "firstCapturedAt": {
            "type": ["string", "null"]
          },
          "lastCapturedAt": {
            "type": ["string", "null"]
          },
          "typeBreakdown": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "topTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            }
          }
        },
        "required": ["noteCount", "slug", "topTags", "typeBreakdown"]
      },
      "TagsResponse": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            }
          }
        },
        "required": ["tags"]
      },
      "SuggestedTopicResponse": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "noteCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["noteCount", "score", "slug"]
      },
      "SuggestionsResponse": {
        "type": "object",
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuggestedTopicResponse"
            }
          }
        },
        "required": ["suggestions"]
      },
      "SourceResponse": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["count", "source"]
      },
      "SourcesResponse": {
        "type": "object",
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceResponse"
            }
          }
        },
        "required": ["sources"]
      },
      "ScopeResponse": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string"
          },
          "noteCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastCapturedAt": {
            "type": ["string", "null"]
          }
        },
        "required": ["noteCount", "scope"]
      },
      "ScopesResponse": {
        "type": "object",
        "properties": {
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScopeResponse"
            }
          }
        },
        "required": ["scopes"]
      },
...
      "InboxNoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "vaultPath": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["capturedAt", "id", "scope", "source", "tags", "title", "type", "vaultPath"]
      },
      "InboxResponse": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InboxNoteResponse"
            }
          }
        },
        "required": ["notes"]
      },
      "DuplicateMatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["id", "scope", "score", "title"]
      },
      "DuplicatesResponse": {
        "type": "object",
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DuplicateMatchResponse"
            }
          }
        },
        "required": ["matches"]
      },
      "AuditResponse": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditRowResponse"
            }
          }
        },
        "required": ["rows"]
      },
      "AuditRowResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "targetId": {
            "type": ["string", "null"]
          },
          "targetKind": {
            "type": ["string", "null"]
          },
          "beforeJson": {
            "type": ["string", "null"]
          },
          "afterJson": {
            "type": ["string", "null"]
          },
          "at": {
            "type": "string"
          }
        },
        "required": ["action", "actor", "at", "id"]
      }
    }
  }
}